% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.R
\name{DeepRegressor}
\alias{DeepRegressor}
\title{Deep regression models}
\usage{
DeepRegressor(
  obj,
  n_layers = 3L,
  verbose = 0,
  ignore_warnings = TRUE,
  custom_metric = NULL,
  predictions = FALSE,
  random_state = 42L,
  regressors = "all",
  ...
)
}
\arguments{
\item{obj}{a model object}

\item{n_layers}{number of hidden layers}

\item{verbose}{monitor progress (\code{0}, default, is false and \code{1} is true)}

\item{ignore_warnings}{print trace when model fitting failed}

\item{custom_metric}{defining a custom metric (default is \code{NULL})}

\item{predictions}{obtain predictions (default is \code{FALSE})}

\item{random_state}{reproducibility seed}

\item{...}{additional parameters to be passed to \code{\link{nnetsauce::CustomRegressor}}}
}
\description{
See also https://techtonique.github.io/nnetsauce/
}
\examples{

X <- MASS::Boston[,-14] # dataset has an ethical problem
y <- MASS::Boston$medv

set.seed(13)
(index_train <- base::sample.int(n = nrow(X),
                                 size = floor(0.8*nrow(X)),
                                 replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]

obj2 <- sklearn$linear_model$ElasticNet()

obj <- DeepRegressor(obj2, n_layers = 3L, n_clusters=2L)
res <- obj$fit(X_train, X_test, y_train, y_test)
print(res[[1]])


}
