% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyRegressor}
\alias{LazyRegressor}
\title{Automated Machine Learning for regression models}
\usage{
LazyRegressor(
  verbose = 0,
  ignore_warnings = TRUE,
  custom_metric = NULL,
  predictions = FALSE,
  random_state = 42L,
  estimators = "all",
  preprocess = FALSE,
  ...
)
}
\arguments{
\item{verbose}{monitor progress (\code{0}, default, is false and \code{1} is true)}

\item{ignore_warnings}{print trace when model fitting failed}

\item{custom_metric}{defining a custom metric (default is \code{NULL})}

\item{predictions}{obtain predictions (default is \code{FALSE})}

\item{random_state}{reproducibility seed}

\item{estimators}{specify regressors to be adjusted (default is 'all')}

\item{preprocess}{preprocessing input covariates (default is FALSE \code{FALSE})}

\item{...}{additional parameters to be passed to \code{\link{nnetsauce::CustomRegressor}}}
}
\description{
See also https://techtonique.github.io/nnetsauce/
}
\examples{

X <- MASS::Boston[,-14] # dataset has an ethical problem
y <- MASS::Boston$medv

set.seed(13)
(index_train <- base::sample.int(n = nrow(X),
                                 size = floor(0.8*nrow(X)),
                                 replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]

obj <- LazyRegressor()
res <- obj$fit(X_train, X_test, y_train, y_test)
print(res[[1]])


}
