% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MTS}
\alias{plot.MTS}
\title{Plot multivariate time series forecast or residuals}
\usage{
\method{plot}{MTS}(x, selected_series, level = 95, ...)
}
\arguments{
\item{x}{result from \code{basicf}, \code{ridge2f} or \code{varf} (multivariate time series forecast)}

\item{selected_series}{name of the time series selected for plotting}

\item{level}{confidence levels for prediction intervals}

\item{...}{additional parameters to be passed to \code{plot} or \code{matplot}}

\item{type}{"pi": basic prediction intervals;
"dist": a distribution of predictions; "sims": the simulations}
}
\description{
Plot multivariate time series forecast or residuals
}
