% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileregressor.R
\name{QuantileRegressor}
\alias{QuantileRegressor}
\title{Model-agnostic quantile regressor}
\usage{
QuantileRegressor(
  obj,
  level = 95,
  scoring = "predictions",
  venv_path = "./venv",
  ...
)
}
\arguments{
\item{obj}{A model object}

\item{level}{The quantile level}

\item{scoring}{The scoring function ("predictions", "residuals", "conformal", "studentized", "conformal-studentized")}

\item{venv_path}{The path to the virtual environment}

\item{...}{Additional arguments}
}
\description{
Model-agnostic quantile regressor
}
\examples{

library(datasets)

X <- as.matrix(mtcars[, -1])
y <- mtcars[, 1]

n <- dim(X)[1]
p <- dim(X)[2]

set.seed(213)
train_index <- sample(x = 1:n, size = floor(0.8*n), replace = TRUE)
test_index <- -train_index

X_train <- as.matrix(mtcars[train_index, -1])
y_train <- mtcars[train_index, 1]
X_test <- as.matrix(mtcars[test_index, -1])
y_test <- mtcars[test_index, 1]

sklearn <- nnetsauce::get_sklearn()
obj <- sklearn$ensemble$RandomForestRegressor()
obj2 <- QuantileRegressor(obj)
obj2$fit(X_train, y_train)
print(obj2$score(X_test, y_test))

}
