% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_forecast}
\alias{to_forecast}
\title{Transform list to forecast or mforecast object}
\usage{
to_forecast(out, start_input, frequency_input)
}
\arguments{
\item{out}{a list, an object constructed by \code{MTS} or \code{DeepMTS}}

\item{start_input}{start of the input time series (as in \code{stats::start})}

\item{frequency_input}{frequency of the input time series (as in \code{stats::frequency})}
}
\value{
an object of class \code{forecast} or \code{mforecast}
}
\description{
Transform list to forecast or mforecast object
}
