% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.R
\name{DeepClassifier}
\alias{DeepClassifier}
\title{Deep classification models}
\usage{
DeepClassifier(obj, n_layers = 3L, venv_path = "./venv", ...)
}
\arguments{
\item{obj}{a model object}

\item{n_layers}{number of hidden layers}

\item{...}{additional parameters to be passed to \code{\link{nnetsauce::CustomClassifier}}}
}
\description{
See also https://techtonique.github.io/nnetsauce/
}
\examples{

library(datasets)

set.seed(123)
X <- as.matrix(iris[, 1:4])
y <- as.integer(iris$Species) - 1L

(index_train <- base::sample.int(n = nrow(X),
                                 size = floor(0.8*nrow(X)),
                                 replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]

obj2 <- sklearn$linear_model$LogisticRegressionCV()

obj <- DeepClassifier(obj2, n_layers = 3L)
res <- obj$fit(X_train, y_train)

# accuracy
print(mean(obj$predict(X_test)==y_test))

}
