% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.R
\name{DeepMTS}
\alias{DeepMTS}
\title{Deep MTS models}
\usage{
DeepMTS(
  obj,
  n_layers = 3L,
  n_hidden_features = 5L,
  activation_name = "relu",
  a = 0.01,
  nodes_sim = "sobol",
  bias = TRUE,
  dropout = 0,
  direct_link = TRUE,
  n_clusters = 2L,
  cluster_encode = TRUE,
  type_clust = "kmeans",
  lags = 1L,
  replications = NULL,
  kernel = NULL,
  agg = "mean",
  seed = 123L,
  backend = c("cpu", "gpu", "tpu"),
  verbose = 0,
  venv_path = "./venv",
  ...
)
}
\arguments{
\item{obj}{a model object}

\item{n_layers}{number of hidden layers}

\item{...}{additional parameters to be passed to \code{\link{nnetsauce::CustomRegressor}}}
}
\description{
See also https://techtonique.github.io/nnetsauce/
}
\examples{

set.seed(123)
X <- matrix(rnorm(300), 100, 3)

obj <- sklearn$linear_model$ElasticNet()
obj2 <- DeepMTS(obj)

obj2$fit(X)
obj2$predict()


}
