% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazydeep.R
\name{LazyDeepClassifier}
\alias{LazyDeepClassifier}
\title{Automated Machine Learning for deep classification models}
\usage{
LazyDeepClassifier(
  n_layers = 3L,
  verbose = 0,
  ignore_warnings = TRUE,
  custom_metric = NULL,
  predictions = FALSE,
  random_state = 42L,
  classifiers = "all",
  preprocess = FALSE,
  ...
)
}
\arguments{
\item{n_layers}{number of hidden layers}

\item{verbose}{monitor progress (\code{0}, default, is false and \code{1} is true)}

\item{ignore_warnings}{print trace when model fitting failed}

\item{custom_metric}{defining a custom metric (default is \code{NULL})}

\item{predictions}{obtain predictions (default is \code{FALSE})}

\item{random_state}{reproducibility seed}

\item{classifiers}{specify classifiers to be adjusted (default is 'all')}

\item{preprocess}{preprocessing input covariates (default is FALSE \code{FALSE})}

\item{...}{additional parameters to be passed to \code{\link{nnetsauce::CustomClassifier}}}
}
\description{
Automated Machine Learning for deep classification models
}
\examples{

library(datasets)

set.seed(123)
X <- as.matrix(iris[, 1:4])
y <- as.integer(iris$Species) - 1L

(index_train <- base::sample.int(n = nrow(X),
                                 size = floor(0.8*nrow(X)),
                                 replace = FALSE))
X_train <- X[index_train, ]
y_train <- y[index_train]
X_test <- X[-index_train, ]
y_test <- y[-index_train]

obj <- LazyDeepClassifier(n_layers = 3L)
res <- obj$fit(X_train, X_test, y_train, y_test)
print(res[[1]])

}
